finish file reorg
