+++ woodwork/serializers/parquet_serializer.py
@@ -1,8 +1,8 @@
 from woodwork.accessor_utils import _is_dask_dataframe, _is_koalas_dataframe
-from woodwork.serializer_utils import clean_latlong
 from woodwork.serializers.serializer_base import (
     PYARROW_IMPORT_ERROR_MESSAGE,
     Serializer,
+    clean_latlong,
 )
 from woodwork.utils import import_or_raise
