+++ woodwork/serializers/serializer_base.py
@@ -5,6 +5,7 @@ import tarfile
 import tempfile
 
 from woodwork.accessor_utils import _is_dask_dataframe, _is_koalas_dataframe
+from woodwork.logical_types import LatLong
 from woodwork.s3_utils import get_transport_params, use_smartopen
 from woodwork.type_sys.utils import _get_ltype_class, _get_specified_ltype_params
 from woodwork.utils import _is_s3, _is_url
@@ -178,3 +179,18 @@ def typing_info_to_dict(dataframe):
         "loading_info": {"table_type": table_type},
         "table_metadata": dataframe.ww.metadata,
     }
+
+
+def clean_latlong(dataframe):
+    """Convert latlong tuples to strings for parquet, arrow and feather file format.
+    Attempting to serialize with tuples present results in an error"""
+    latlong_columns = [
+        col_name
+        for col_name, col in dataframe.ww.columns.items()
+        if _get_ltype_class(col.logical_type) == LatLong
+    ]
+    if len(latlong_columns) > 0:
+        dataframe = dataframe.ww.copy()
+        dataframe[latlong_columns] = dataframe[latlong_columns].astype(str)
+
+    return dataframe
